* 'master' of https://github.com/openMSX/openMSX: (1200 commits)
  Allow user to control all disk drives/cart slots in OSD menu.
  Move pause_on_lost_focus from c++ to Tcl
  Avoid (harmless) error when closing the OSD menu via clicking the button in the top-left corner
  In get_screen fetch the character under the cursor.
  small cleanup (leftover from 4c86ae715ab)
  Fix mirroring of SFG-1, it's all over the slot.
  Document that CX5 is identical to CX5F, except keyboard.
  Deactivate the input layer as late as possible.
  Added bindings for both COPY and PASTE...
  Decode all output of spawned commands as UTF-8
  Yet another fix for writing to encoded text files
  Another fix for writing to encoded text files
  Fix writing to encoded text files
  Use print function syntax
  Use next() function instead of method
  Use xrange() instead of range()
  Modernize writing of text files
  Modernize writing of binary file
  Modernize use of fdopen()
  Use modern syntax to open tar files
  ...

+++ src/sound/NullSoundDriver.cc
@@ -20,7 +20,7 @@ unsigned NullSoundDriver::getSamples() const
 	return 0;
 }
 
-void NullSoundDriver::uploadBuffer(short* /*buffer*/, unsigned /*len*/)
+void NullSoundDriver::uploadBuffer(float* /*buffer*/, unsigned /*len*/)
 {
 }
