+++ src/video/scalers/Simple2xScaler.hh
@@ -39,9 +39,9 @@ private:
 	void drawScanline(const Pixel* in1, const Pixel* in2, Pixel* out,
 	                  int factor, unsigned dstWidth);
 	void blur1on2(const Pixel* pIn, Pixel* pOut, unsigned alpha,
-	              unsigned long srcWidth);
+	              size_t srcWidth);
 	void blur1on1(const Pixel* pIn, Pixel* pOut, unsigned alpha,
-	              unsigned long srcWidth);
+	              size_t srcWidth);
 
 	RenderSettings& settings;
 	PixelOperations<Pixel> pixelOps;
